Stop Inlining Critical CSS
Since the main CSS bundle is still being loaded in a blocking way,
inlining critical CSS is actually detrimental to the first page
load experience (because the CSS isn't extracted from the bundle
and must be parsed twice). The whole bundle needs to be loaded
for SPA page transitions to work right anyways, and it really
isn't very large. Plus, with HTTP2 push the external CSS file
can be pre-sent.